[add] 그룹 리스트 정보 반환 이동
